Feature/wmt 475/workload report (#23)
* Fixing issues with schema scripts and added workload report
- Workload Report domain and data insert logic
- Separated out permission statements which are dependant on schema existing

* Ignore eslint warnings

* Moved persistance layer from PostgreSQL to Azure SQL
- Updated permissions and user creation scripts
- Updated migration scripts to be ASQL compatible

+++ migrations/app/20170508150632_app_workload.js
@@ -3,7 +3,7 @@ exports.up = function (knex, Promise) {
   return knex.schema.withSchema('app').createTable('workload', function (table) {
     table.increments('id')
     table.integer('workload_owner_id').unsigned().references('workload_owner.id')
-    table.bigInteger('workload_points_calc_id').unsigned().references('workload_points_calculations.id')
+    table.integer('workload_points_calc_id').unsigned().references('workload_points_calculations.id')
     table.integer('total_cases').unsigned().notNullable()
     table.integer('total_cases_inactive').unsigned().notNullable()
     table.integer('monthly_sdrs').unsigned().notNullable()
