+++ migrations/app/20170508160700_app_license_tiers.js
@@ -2,10 +2,10 @@
 exports.up = function (knex, Promise) {
   return knex.schema.createTable('license_tiers', function (table) {
     table.increments('id')
-    table.bigInteger('workload_id').unsigned().notNullable().references('workload.id')
+    table.integer('workload_id').unsigned().notNullable().references('workload.id')
     table.integer('tier_type').unsigned()
     table.string('case_type', 5)
-    table.bigInteger('points').unsigned()
+    table.integer('points').unsigned()
   }).catch(function (error) {
     console.log(error)
     throw error
